refactor(rendering): extract 4k quality helpers
